backend/requirements.txt
fastapi
uvicorn[standard]
dspy-ai
langgraph
playwright
python-dotenv
pytest
openpyxl
python-docx
pymupdf
xlsxwriter
pandas
pyautogui
pygetwindow
pyperclip
AppOpener
send2trash
docxtpl
pdfplumber
beautifulsoup4
lxml
reportlab
openai
yfinance
faster-whisper
requests
pyinstaller
opencv-python
pytesseract
pillow
groq
